New upstream version 2.9.0

+++ log4j-core/src/main/java/org/apache/logging/log4j/core/appender/rolling/RollingRandomAccessFileManager.java
@@ -47,12 +47,28 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
     private RandomAccessFile randomAccessFile;
     private final ThreadLocal<Boolean> isEndOfBatch = new ThreadLocal<>();
 
+    @Deprecated
     public RollingRandomAccessFileManager(final LoggerContext loggerContext, final RandomAccessFile raf,
             final String fileName, final String pattern, final OutputStream os, final boolean append,
             final boolean immediateFlush, final int bufferSize, final long size, final long time,
             final TriggeringPolicy policy, final RolloverStrategy strategy, final String advertiseURI,
             final Layout<? extends Serializable> layout, final boolean writeHeader) {
+        this(loggerContext, raf, fileName, pattern, os, append, immediateFlush, bufferSize, size, time, policy, strategy, advertiseURI,
+               layout, null, null, null, writeHeader);
+    }
+
+    /**
+     * @since 2.8.3
+     */
+    public RollingRandomAccessFileManager(final LoggerContext loggerContext, final RandomAccessFile raf,
+            final String fileName, final String pattern, final OutputStream os, final boolean append,
+            final boolean immediateFlush, final int bufferSize, final long size, final long time,
+            final TriggeringPolicy policy, final RolloverStrategy strategy, final String advertiseURI,
+            final Layout<? extends Serializable> layout,
+            final String filePermissions, final String fileOwner, final String fileGroup,
+            final boolean writeHeader) {
         super(loggerContext, fileName, pattern, os, append, false, size, time, policy, strategy, advertiseURI, layout,
+                filePermissions, fileOwner, fileGroup,
                 writeHeader, ByteBuffer.wrap(new byte[bufferSize]));
         this.randomAccessFile = raf;
         isEndOfBatch.set(Boolean.FALSE);
@@ -83,9 +99,16 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
     public static RollingRandomAccessFileManager getRollingRandomAccessFileManager(final String fileName,
             final String filePattern, final boolean isAppend, final boolean immediateFlush, final int bufferSize,
             final TriggeringPolicy policy, final RolloverStrategy strategy, final String advertiseURI,
-            final Layout<? extends Serializable> layout, final Configuration configuration) {
-        return (RollingRandomAccessFileManager) getManager(fileName, new FactoryData(filePattern, isAppend,
-                immediateFlush, bufferSize, policy, strategy, advertiseURI, layout, configuration), FACTORY);
+            final Layout<? extends Serializable> layout, final String filePermissions, final String fileOwner, final String fileGroup,
+            final Configuration configuration) {
+        if (strategy instanceof DirectWriteRolloverStrategy && fileName != null) {
+            LOGGER.error("The fileName attribute must not be specified with the DirectWriteRolloverStrategy");
+            return null;
+        }
+        final String name = fileName == null ? filePattern : fileName;
+        return narrow(RollingRandomAccessFileManager.class, getManager(name, new FactoryData(fileName, filePattern, isAppend,
+                immediateFlush, bufferSize, policy, strategy, advertiseURI, layout,
+                filePermissions, fileOwner, fileGroup, configuration), FACTORY));
     }
 
     public Boolean isEndOfBatch() {
@@ -106,6 +129,12 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
     @Override
     protected synchronized void writeToDestination(final byte[] bytes, final int offset, final int length) {
         try {
+            if (randomAccessFile == null) {
+                String fileName = getFileName();
+                File file = new File(fileName);
+                FileUtils.makeParentDirs(file);
+                createFileAfterRollover(fileName);
+            }
             randomAccessFile.write(bytes, offset, length);
             size += length;
         } catch (final IOException ex) {
@@ -116,7 +145,11 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
 
     @Override
     protected void createFileAfterRollover() throws IOException {
-        this.randomAccessFile = new RandomAccessFile(getFileName(), "rw");
+        createFileAfterRollover(getFileName());
+    }
+
+    private void createFileAfterRollover(String fileName) throws IOException {
+        this.randomAccessFile = new RandomAccessFile(fileName, "rw");
         if (isAppend()) {
             randomAccessFile.seek(randomAccessFile.length());
         }
@@ -165,41 +198,52 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
          */
         @Override
         public RollingRandomAccessFileManager createManager(final String name, final FactoryData data) {
-            final File file = new File(name);
-
-            if (!data.append) {
-                file.delete();
-            }
-            final long size = data.append ? file.length() : 0;
-            final long time = file.exists() ? file.lastModified() : System.currentTimeMillis();
-
-            final boolean writeHeader = !data.append || !file.exists();
+            File file = null;
+            long size = 0;
+            long time = System.currentTimeMillis();
             RandomAccessFile raf = null;
-            try {
-                FileUtils.makeParentDirs(file);
-                raf = new RandomAccessFile(name, "rw");
-                if (data.append) {
-                    final long length = raf.length();
-                    LOGGER.trace("RandomAccessFile {} seek to {}", name, length);
-                    raf.seek(length);
-                } else {
-                    LOGGER.trace("RandomAccessFile {} set length to 0", name);
-                    raf.setLength(0);
+            if (data.fileName != null) {
+                file = new File(name);
+
+                if (!data.append) {
+                    file.delete();
                 }
-                return new RollingRandomAccessFileManager(data.getLoggerContext(), raf, name, data.pattern,
-                        NullOutputStream.getInstance(), data.append, data.immediateFlush, data.bufferSize, size, time, data.policy,
-                        data.strategy, data.advertiseURI, data.layout, writeHeader);
-            } catch (final IOException ex) {
-                LOGGER.error("Cannot access RandomAccessFile " + ex, ex);
-                if (raf != null) {
-                    try {
-                        raf.close();
-                    } catch (final IOException e) {
-                        LOGGER.error("Cannot close RandomAccessFile {}", name, e);
+                size = data.append ? file.length() : 0;
+                if (file.exists()) {
+                    time = file.lastModified();
+                }
+                try {
+                    FileUtils.makeParentDirs(file);
+                    raf = new RandomAccessFile(name, "rw");
+                    if (data.append) {
+                        final long length = raf.length();
+                        LOGGER.trace("RandomAccessFile {} seek to {}", name, length);
+                        raf.seek(length);
+                    } else {
+                        LOGGER.trace("RandomAccessFile {} set length to 0", name);
+                        raf.setLength(0);
                     }
+                } catch (final IOException ex) {
+                    LOGGER.error("Cannot access RandomAccessFile " + ex, ex);
+                    if (raf != null) {
+                        try {
+                            raf.close();
+                        } catch (final IOException e) {
+                            LOGGER.error("Cannot close RandomAccessFile {}", name, e);
+                        }
+                    }
+                    return null;
                 }
             }
-            return null;
+            final boolean writeHeader = !data.append || file == null || !file.exists();
+
+            final RollingRandomAccessFileManager rrm = new RollingRandomAccessFileManager(data.getLoggerContext(), raf, name, data.pattern,
+                    NullOutputStream.getInstance(), data.append, data.immediateFlush, data.bufferSize, size, time, data.policy,
+                    data.strategy, data.advertiseURI, data.layout, data.filePermissions, data.fileOwner, data.fileGroup, writeHeader);
+            if (rrm.isAttributeViewEnabled()) {
+                rrm.defineAttributeView(file.toPath());
+            }
+            return rrm;
         }
     }
 
@@ -207,6 +251,7 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
      * Factory data.
      */
     private static class FactoryData extends ConfigurationFactoryData {
+        private final String fileName;
         private final String pattern;
         private final boolean append;
         private final boolean immediateFlush;
@@ -215,10 +260,14 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
         private final RolloverStrategy strategy;
         private final String advertiseURI;
         private final Layout<? extends Serializable> layout;
+        private final String filePermissions;
+        private final String fileOwner;
+        private final String fileGroup;
 
         /**
          * Create the data for the factory.
          *
+         * @param fileName The file name.
          * @param pattern The pattern.
          * @param append The append flag.
          * @param immediateFlush
@@ -227,12 +276,18 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
          * @param strategy
          * @param advertiseURI
          * @param layout
+         * @param filePermissions File permissions
+         * @param fileOwner File owner
+         * @param fileGroup File group
          * @param configuration
          */
-        public FactoryData(final String pattern, final boolean append, final boolean immediateFlush,
+        public FactoryData(final String fileName, final String pattern, final boolean append, final boolean immediateFlush,
                 final int bufferSize, final TriggeringPolicy policy, final RolloverStrategy strategy,
-                final String advertiseURI, final Layout<? extends Serializable> layout, final Configuration configuration) {
+                final String advertiseURI, final Layout<? extends Serializable> layout,
+                final String filePermissions, final String fileOwner, final String fileGroup,
+                final Configuration configuration) {
             super(configuration);
+            this.fileName = fileName;
             this.pattern = pattern;
             this.append = append;
             this.immediateFlush = immediateFlush;
@@ -241,6 +296,9 @@ public class RollingRandomAccessFileManager extends RollingFileManager {
             this.strategy = strategy;
             this.advertiseURI = advertiseURI;
             this.layout = layout;
+            this.filePermissions = filePermissions;
+            this.fileOwner = fileOwner;
+            this.fileGroup = fileGroup;
         }
 
         public TriggeringPolicy getTriggeringPolicy()
